[refactor] :recycle: adapt site to new bff folders structures
