+++ src/sections/banking/view/overview-banking-view.tsx
@@ -9,7 +9,7 @@ import Grid from '@mui/material/Unstable_Grid2'
 import { useTranslate } from 'src/locales'
 import { AuthUserType } from 'src/auth/types'
 import { useAuthContext } from 'src/auth/hooks'
-import { useGetWalletBalance, useGetWalletTransactions } from 'src/app/api/_hooks'
+import { useGetWalletBalance, useGetWalletTransactions } from 'src/app/api/hooks'
 
 import { useSettingsContext } from 'src/components/settings'
