+++ src/sections/nfts/view/nfts-view.tsx
@@ -7,7 +7,7 @@ import Container from '@mui/material/Container'
 import { useTranslate } from 'src/locales'
 import { AuthUserType } from 'src/auth/types'
 import { useAuthContext } from 'src/auth/hooks'
-import { useGetWalletNfts } from 'src/app/api/_hooks/'
+import { useGetWalletNfts } from 'src/app/api/hooks/'
 
 import EmptyContent from 'src/components/empty-content'
 import { useSettingsContext } from 'src/components/settings'
